+++ docker-entrypoint.sh
@@ -2,8 +2,8 @@
 
 set -eu
 
-if [ "$1" = 'runwsgi' ]; then
-	exec uwsgi --ini /minime/uwsgi.ini --uid=1000 --gid=2000
+if [ "$1" = 'runproduction' ]; then
+	exec python /minime/manage.py collectstatic --noinput && gunicorn app.wsgi -d 0.0.0.0:8000
 elif [ "$1" = 'runserver' ]; then
 	exec python /minime/manage.py runserver 0.0.0.0:8000
 elif [ "$1" = 'test' ]; then